initial commit
this is a group project which is held nation-wide in Turkey, we aimed to help people who suffer from back diseases e.g. scoliosis and others. we tried to achieve a solution with using flex sensors who collect data on a patient's back and processing the data on a arduino